Simplify and actually make the configmap reconciler work.

+++ pkg/configmaps/lib.go
@@ -3,16 +3,13 @@ package configmaps
 import (
 	"context"
 	"crypto/md5"
-	"errors"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 
-	"github.com/operator-framework/operator-sdk/pkg/predicate"
 	corev1 "k8s.io/api/core/v1"
-	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
@@ -31,12 +28,12 @@ type ConfigMapReconcilerConfig struct {
 }
 
 type configMapReconciler struct {
-	client    client.Client
-	config    ConfigMapReconcilerConfig
-	selector  labels.Selector
-	baseDir   string
-	namespace string
-	fileCache map[types.NamespacedName]configFiles
+	client       client.Client
+	clientConfig *rest.Config
+	config       ConfigMapReconcilerConfig
+	selector     labels.Selector
+	baseDir      string
+	namespace    string
 }
 
 // configFiles is a map where keys are the names of the files and values are digests of their content
@@ -50,39 +47,52 @@ func New(mgr manager.Manager, config ConfigMapReconcilerConfig) error {
 	}
 
 	r := &configMapReconciler{
-		client:    mgr.GetClient(),
-		config:    config,
-		selector:  lbls.AsSelector(),
-		fileCache: make(map[types.NamespacedName]configFiles),
+		client:       mgr.GetClient(),
+		clientConfig: mgr.GetConfig(),
+		config:       config,
+		selector:     lbls.AsSelector(),
 	}
 
 	// register the controller with the manager
 	bld := builder.ControllerManagedBy(mgr)
 	bld.Named("config-bump")
-	bld.WithEventFilter(predicate.ResourceFilterPredicate{Selector: r.selector})
+	bld.ForType(&corev1.ConfigMap{})
+	// note that we do NOT set up the filter to only included the labeled config maps.
+	// That way we would never see the events about deleted config maps or config maps from which
+	// the label has been removed
+	//bld.WithEventFilter(predicate.ResourceFilterPredicate{Selector: r.selector})
 	if err = bld.Complete(r); err != nil {
 		return err
 	}
 
-	r.Initialize()
+	r.sync(false)
 
 	return nil
 }
 
-// Initialize performs an initial sync of the local set of files with the configured config maps
-// It also initializes the watches to initialize the reconciliation loop
-func (c *configMapReconciler) Initialize() error {
+// sync performs the sync of the local set of files with the configured config maps
+func (c *configMapReconciler) sync(managerRunning bool) error {
+	var cl client.Client
+	if managerRunning {
+		cl = c.client
+	} else {
+		x, err := client.New(c.clientConfig, client.Options{})
+		if err != nil {
+			return err
+		}
+		cl = x
+	}
+
 	list := &corev1.ConfigMapList{}
 	opts := []client.ListOption{
 		client.InNamespace(c.config.Namespace),
 		client.MatchingLabelsSelector{Selector: c.selector},
 	}
 
-	if err := c.client.List(context.TODO(), list, opts...); err != nil {
+	if err := cl.List(context.TODO(), list, opts...); err != nil {
 		return err
 	}
 
-	c.fileCache = make(map[types.NamespacedName]configFiles)
 	processedFiles := make([]string, 0, 8)
 
 	for _, cm := range list.Items {
@@ -127,6 +137,18 @@ func (c *configMapReconciler) Initialize() error {
 	for _, f := range files {
 		if !f.IsDir() {
 			path := filepath.Join(c.config.BaseDir, f.Name())
+			found := false
+			for _, v := range processedFiles {
+				if v == path {
+					found = true
+					break
+				}
+			}
+
+			if found {
+				continue
+			}
+
 			if err := os.Remove(path); err != nil {
 				return err
 			}
@@ -138,85 +160,6 @@ func (c *configMapReconciler) Initialize() error {
 
 // Reconcile handles the changes in the configured config maps
 func (c *configMapReconciler) Reconcile(r reconcile.Request) (reconcile.Result, error) {
-	cm := &corev1.ConfigMap{}
-	err := c.client.Get(context.TODO(), r.NamespacedName, cm)
-	if err != nil && !kerrors.IsNotFound(err) {
-		return reconcile.Result{}, errors.New("error while retrieving object")
-	}
-
-	// map of file names and their content as found in the config map
-	toCreate := make(map[string][]byte)
-	toUpdate := make(map[string][]byte)
-	toDelete := make([]string, 0, 8)
-
-	cachedFiles, cached := c.fileCache[r.NamespacedName]
-	if err == nil {
-		if cached {
-			for name, data := range cm.Data {
-				if existingHash, ok := cachedFiles[name]; ok {
-					// we already know about this file - let's see if it changed
-					hash := md5.Sum([]byte(data))
-
-					if hash != existingHash {
-						toUpdate[name] = []byte(data)
-					}
-				} else {
-					// our cache doesn't contain this file...
-					toCreate[name] = []byte(data)
-				}
-			}
-		} else {
-			for name, data := range cm.Data {
-				toCreate[name] = []byte(data)
-			}
-		}
-	} else {
-		// not found - one of the config maps has been deleted
-		if cached {
-			for name := range cachedFiles {
-				toDelete = append(toDelete, name)
-			}
-		} else {
-			// we got report about a deletion of a config map that was not in our cache.
-			// This calls for a full scan
-			if err := c.Initialize(); err != nil {
-				return reconcile.Result{}, err
-			}
-		}
-	}
-
-	for name, data := range toCreate {
-		f, err := os.Create(filepath.Join(c.config.BaseDir, name))
-		if err != nil {
-			return reconcile.Result{}, err
-		}
-
-		defer f.Close()
-
-		if _, err = f.Write(data); err != nil {
-			return reconcile.Result{}, err
-		}
-	}
-
-	for name, data := range toUpdate {
-		f, err := os.Create(filepath.Join(c.config.BaseDir, name))
-		if err != nil {
-			return reconcile.Result{}, err
-		}
-
-		defer f.Close()
-
-		if _, err = f.Write(data); err != nil {
-			return reconcile.Result{}, err
-		}
-	}
-
-	for _, name := range toDelete {
-		err := os.Remove(filepath.Join(c.config.BaseDir, name))
-		if err != nil {
-			return reconcile.Result{}, err
-		}
-	}
-
-	return reconcile.Result{}, nil
+	err := c.sync(true)
+	return reconcile.Result{}, err
 }
